# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "XSocial Laravel API",
        "description": "✊✊✊",
        "contact": {
            "email": "ngoctoan06011998@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/follows": {
            "get": {
                "tags": [
                    "Follows"
                ],
                "summary": "follows.index",
                "operationId": "3791fc7bdb235572ab9b037614de4a21",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "filter[user_id]",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "filter[follower_user_id]",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "user|follower_user|following_user",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Follows"
                ],
                "summary": "follows.store",
                "description": "Start follow an user",
                "operationId": "a7ac4dd4f950be0c69089ed1fc2263b6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "follower_user_id"
                                ],
                                "properties": {
                                    "follower_user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/follows/{follower_user_id}": {
            "delete": {
                "tags": [
                    "Follows"
                ],
                "summary": "follows.destroy",
                "description": "Unfollow an user",
                "operationId": "8ae58077e2313306877bd794832e301f",
                "parameters": [
                    {
                        "name": "follower_user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "messages.index",
                "operationId": "9d5effa2cb648df85be0ea9741ffb457",
                "parameters": [
                    {
                        "name": "topic_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "favorites|reply",
                        "required": false
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "-id, -created_at",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "messages.store",
                "operationId": "db79bb7bcadfdb708e9f6c38ac928f4a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "topic_id",
                                    "body",
                                    "media_ids"
                                ],
                                "properties": {
                                    "topic_id": {
                                        "type": "integer"
                                    },
                                    "body": {
                                        "type": "string"
                                    },
                                    "reply_id": {
                                        "type": "integer"
                                    },
                                    "media_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/messages/{id}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "messages.show",
                "operationId": "8938bb52056aef543d340164d37f9562",
                "parameters": [
                    {
                        "name": "topic_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "favorites|reply",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Bearer {token}"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}